Assemble proteoforms and calculate their masses.

This function combines a set of proteins with a set of PTM compositions ("modcoms") and calculates the masses of the resulting proteoforms (i.e., the different molecular forms in which the protein product of a single gene can be found due to changes introduced by posttranslational modifications).

assemble_proteoforms(p, m, mass_set = "average")

Arguments

p
A protein specification as returned by define_proteins().

m
A modcom specification as returned by define_ptm_compositions().

mass_set
Atomic masses that should be used for mass calculation.

Value

A data frame containing \(p \times m\) proteoforms, obtained by combining information on \(p\) proteins (i.e., number of rows in p) with information on \(m\) PTM compositions (i.e., number of rows in m). This data frame comprises two columns derived from the input data frames (protein_name and modcom_name, respectively) and two new columns formula (overall molecular formula) and mass (calculated proteoform masses).

See also

Examples

proteins <- define_proteins(
  system.file("extdata", "mab_sequence.fasta", package = "fragquaxi"),
  .disulfides = 16
)

modcoms <- define_ptm_compositions(sample_modcoms)

assemble_proteoforms(proteins, modcoms)

#> # A tibble: 6 x 4
#>   protein_name modcom_name formula                         mass
#>          <int> <chr>       <mol>                          <dbl>
#> 1            1 G0F/G0      C6570 H10124 N1714 O2088 S44 147942.
#> 2            1 G0F/G0F     C6576 H10134 N1714 O2092 S44 148088.
#> 3            1 G0F/G1F     C6582 H10144 N1714 O2097 S44 148250.
#> 4            1 G1F/G1F     C6588 H10154 N1714 O2102 S44 148412.
#> 5            1 G1F/G2F     C6594 H10164 N1714 O2107 S44 148574.
#> 6            1 G2F/G2F     C6600 H10174 N1714 O2112 S44 148737.