Average and monoisotopic atomic masses.

A dataset containing average and monoisotopic masses for the chemical elements. Masses are missing in case of elements for which no stable isotope is known.

atomic_masses

Format

A data frame with 118 rows and 6 variables:

element

chemical symbol

name

IUPAC name

Z

atomic number (number of protons in the nucleus)

average

average atomic mass as defined by IUPAC

A

mass number (total number of protons and neutrons in the nucleus) of the most abundant isotope

monoisotopic

atomic mass of the most abundant isotope

Source

Average masses: Atomic weights of the elements 2013, Tables 2 and 3 (https://doi.org/10.1515/pac-2015-0305)

Isotope abundances: Isotopic compositions of the elements 2013, Table 1 (https://doi.org/10.1515/pac-2015-0503)

Monoisotopic masses: The AME16 Atomic Mass Evaluation (https://www-nds.iaea.org/amdc/ame2016/mass16.txt)