Calculate mass-to-charge ratios.

The function is vectorized over z and thus may calculate several \(m/z\) values for a single mass.

charge(m, z, charge_agent_mass = NULL)

Arguments

m
Mass.

z
Vector of charges.

charge_agent_mass
Mass of the charge carrier. If NULL, use the average IUPAC mass of the proton.

Value

A named vector of mass-to-charge ratios; names correspond to the values of z, converted to character.

Examples

charge(150000, z = 24)

#>       24 
#> 6251.008 

charge(150000, z = 20:24)

#>       20       21       22       23       24 
#> 7501.008 7143.865 6819.190 6522.747 6251.008 

charge(150000, z = 20:24, charge_agent_mass = 3)

#>       20       21       22       23       24 
#> 7503.000 7145.857 6821.182 6524.739 6253.000