Coerce quantification results to a data frame.

This function converts quantification results obtained from quantify_ions() to a nested data frame, which provides data on abundances and extracted ion currents (XIC) for each quantified ion.

as_tibble.quaxi(x, abundance_col = "abundance_data", xic_col = "xic_data", ...)

Arguments

x
A quaxi object that should be coerced to a data frame.

abundance_col
Name of the column that will contain abundance data. If NULL, do not include abundance data in the output.

xic_col
Name of the column that will contain XIC data. If NULL, do not include XIC data in the output.

...
Other arguments passed on to individual methods.

Value

The data frame stored in x$ions without column ion_id (i.e., a data frame with one ion per row) and up to two nested columns containing data on abundances and XICs, respectively.

The former column is labeled according to the value of abundance_col and contains the variables rt_min, rt_max, scan_min, scan_max, and abundance.

The latter column is labeled according to the value of xic_col and contains the variables scan (scan index), rt (corresponding retention time), and int (intensity).

Examples

library(tibble)

ms_data <- mzR::openMSfile(
  system.file("extdata", "mzml", "mab1.mzML", package = "fragquaxi")
)
proteoforms <- tribble(
  ~name,       ~Hex, ~HexNAc, ~Fuc,
  "G0F/G0",       6,       8,    1,
  "G0F/G0F",      6,       8,    2,
  "G0F/G1F",      7,       8,    2,
  "G1F/G1F",      8,       8,    2,
  "G1F/G2F",      9,       8,    2,
  "G2F/G2F",     10,       8,    2,
)

ions <-
  proteoforms %>%
  calculate_proteoform_masses("C6464 H9950 N1706 O2014 S44") %>%
  ionize(36L:40L)

abundances <- quantify_ions(ms_data, ions, c(300, 350))

# by default, include data on abundances and XICs in the output
as_tibble(abundances)

#> # A tibble: 30 x 11
#>    name    Hex HexNAc   Fuc   mass     z    mz mz_min mz_max abundance_data
#>    <chr> <dbl>  <dbl> <dbl>  <dbl> <int> <dbl>  <dbl>  <dbl> <list>        
#>  1 G0F/…     6      8     1 1.48e5    36 4111.  4110.  4111. <tibble [1 × …
#>  2 G0F/…     6      8     1 1.48e5    37 3999.  3999.  4000. <tibble [1 × …
#>  3 G0F/…     6      8     1 1.48e5    38 3894.  3893.  3895. <tibble [1 × …
#>  4 G0F/…     6      8     1 1.48e5    39 3794.  3794.  3795. <tibble [1 × …
#>  5 G0F/…     6      8     1 1.48e5    40 3700.  3699.  3700. <tibble [1 × …
#>  6 G0F/…     6      8     2 1.48e5    36 4115.  4114.  4115. <tibble [1 × …
#>  7 G0F/…     6      8     2 1.48e5    37 4003.  4003.  4004. <tibble [1 × …
#>  8 G0F/…     6      8     2 1.48e5    38 3898.  3897.  3899. <tibble [1 × …
#>  9 G0F/…     6      8     2 1.48e5    39 3798.  3797.  3799. <tibble [1 × …
#> 10 G0F/…     6      8     2 1.48e5    40 3703.  3702.  3704. <tibble [1 × …
#> # … with 20 more rows, and 1 more variable: xic_data <list>

# rename the column containing abundance data
# (and omit several columns for readability)
as_tibble(abundances, abundance_col = "abundances") %>%
  dplyr::select(-Hex:-Fuc, -mz_min, -mz_max)

#> # A tibble: 30 x 6
#>    name       mass     z    mz abundances       xic_data          
#>    <chr>     <dbl> <int> <dbl> <list>           <list>            
#>  1 G0F/G0  147942.    36 4111. <tibble [1 × 5]> <tibble [352 × 3]>
#>  2 G0F/G0  147942.    37 3999. <tibble [1 × 5]> <tibble [352 × 3]>
#>  3 G0F/G0  147942.    38 3894. <tibble [1 × 5]> <tibble [352 × 3]>
#>  4 G0F/G0  147942.    39 3794. <tibble [1 × 5]> <tibble [352 × 3]>
#>  5 G0F/G0  147942.    40 3700. <tibble [1 × 5]> <tibble [352 × 3]>
#>  6 G0F/G0F 148088.    36 4115. <tibble [1 × 5]> <tibble [352 × 3]>
#>  7 G0F/G0F 148088.    37 4003. <tibble [1 × 5]> <tibble [352 × 3]>
#>  8 G0F/G0F 148088.    38 3898. <tibble [1 × 5]> <tibble [352 × 3]>
#>  9 G0F/G0F 148088.    39 3798. <tibble [1 × 5]> <tibble [352 × 3]>
#> 10 G0F/G0F 148088.    40 3703. <tibble [1 × 5]> <tibble [352 × 3]>
#> # … with 20 more rows

# omit XIC data in the output
as_tibble(abundances, xic_col = NULL) %>%
  dplyr::select(-Hex:-Fuc, -mz_min, -mz_max)

#> # A tibble: 30 x 5
#>    name       mass     z    mz abundance_data  
#>    <chr>     <dbl> <int> <dbl> <list>          
#>  1 G0F/G0  147942.    36 4111. <tibble [1 × 5]>
#>  2 G0F/G0  147942.    37 3999. <tibble [1 × 5]>
#>  3 G0F/G0  147942.    38 3894. <tibble [1 × 5]>
#>  4 G0F/G0  147942.    39 3794. <tibble [1 × 5]>
#>  5 G0F/G0  147942.    40 3700. <tibble [1 × 5]>
#>  6 G0F/G0F 148088.    36 4115. <tibble [1 × 5]>
#>  7 G0F/G0F 148088.    37 4003. <tibble [1 × 5]>
#>  8 G0F/G0F 148088.    38 3898. <tibble [1 × 5]>
#>  9 G0F/G0F 148088.    39 3798. <tibble [1 × 5]>
#> 10 G0F/G0F 148088.    40 3703. <tibble [1 × 5]>
#> # … with 20 more rows